--- a/doc/utils/sphinx-config/false_positives.txt
+++ b/doc/utils/sphinx-config/false_positives.txt
@@ -3191,6 +3191,7 @@ Souza
 sp
 spacings
 Spearot
+specieslist
 specular
 spellcheck
 Spellmeyer