--- a/src/Depend.sh
+++ b/src/Depend.sh
@@ -121,6 +121,7 @@ fi
 if (test $1 = "MANYBODY") then
   depend ATC
   depend GPU
+  depend INTEL
   depend KOKKOS
   depend OPT
   depend QEQ