--- a/src/OPENMP/npair_halffull_newtoff_trim_omp.cpp
+++ b/src/OPENMP/npair_halffull_newtoff_trim_omp.cpp
@@ -87,6 +87,8 @@ void NPairHalffullNewtoffTrimOmp::build(NeighList *list)
       joriginal = jlist[jj];
       j = joriginal & NEIGHMASK;
 
+      // trim to shorter cutoff
+
       delx = xtmp - x[j][0];
       dely = ytmp - x[j][1];
       delz = ztmp - x[j][2];