… doc/src/neighbor.rst
… doc/src/pair_modify.rst
… src/KOKKOS/npair_halffull_kokkos.h
… src/KOKKOS/npair_trim_kokkos.cpp
… src/OPENMP/npair_halffull_newtoff_trim_omp.h
… src/OPENMP/npair_halffull_newton_trim_omp.h
… src/OPENMP/npair_trim_omp.cpp
… src/OPENMP/npair_trim_omp.h
… src/neigh_list.cpp
… src/neigh_list.h
… src/neigh_request.cpp
… src/neigh_request.h
… src/neighbor.cpp
… src/neighbor.h
… src/npair_halffull_newtoff_trim.cpp
… src/npair_halffull_newtoff_trim.h
… src/npair_halffull_newton_trim.cpp
… src/npair_halffull_newton_trim.h
… src/npair_trim.cpp
… src/npair_trim.h
… src/pair.cpp
… src/pair.h
… src/pair_hybrid.cpp
… src/pair_hybrid.h
… src/pair_hybrid_scaled.cpp