…with pair hybrid/overlay

**Summary**

This PR adds a `pair_modify` option: `neigh/trim` to allow separate cutoffs for each neigh list with `pair hybrid/overlay` and also adds "trim" methods for building a neighbor list with a shorter cutoff from one with a longer cutoff. This can give speedup when the cutoffs are significantly different.

For example, consider the following input command:

```
pair_style      hybrid/overlay lj/cut 2.5 coul/cut 5.0
pair_coeff      1 1 lj/cut 1.0 1.0
pair_coeff      1 1 coul/cut
```

With `pair_modify neigh/trim no` the code gives the same behavior as before:

```
  master list distance cutoff = 5.3
  ghost atom cutoff = 5.3
  binsize = 2.65, bins = 13 13 13
  2 neighbor lists, perpetual/occasional/extra = 2 0 0
  (1) pair lj/cut, perpetual
      attributes: half, newton on
      pair build: half/bin/atomonly/newton
      stencil: half/bin/3d
      bin: standard
  (2) pair coul/cut, perpetual, copy from (1)
      attributes: half, newton on
      pair build: copy
      stencil: none
      bin: none
```

While with `pair_modify neigh/trim yes` the code builds a separate neigh list for `lj/cut` by trimming the neigh list built for `coul/cut`:

```
  master list distance cutoff = 5.3
  ghost atom cutoff = 5.3
  binsize = 2.65, bins = 13 13 13
  2 neighbor lists, perpetual/occasional/extra = 2 0 0
  (1) pair lj/cut, perpetual, trim from (2)
      attributes: half, newton on, cut 2.8
      pair build: trim
      stencil: none
      bin: none
  (2) pair coul/cut, perpetual
      attributes: half, newton on
      pair build: half/bin/atomonly/newton
      stencil: half/bin/3d
      bin: standard
```



**Related Issue(s)**

None

**Author(s)**

Stan Moore (SNL), with helpful discussions from Steve Plimpton (SNL)

**Licensing**

By submitting this pull request, I agree, that my contribution will be included in LAMMPS and redistributed under either the GNU General Public License version 2 (GPL v2) or the GNU Lesser General Public License version 2.1 (LGPL v2.1).

**Backward Compatibility**

Yes

**Implementation Notes**


**Post Submission Checklist**

- [x] The feature or features in this pull request is complete
- [x] Licensing information is complete
- [x] Corresponding author information is complete
- [x] The source code follows the LAMMPS formatting guidelines
- [x] Suitable new documentation files and/or updates to the existing docs are included
- [ ] The added/updated documentation is integrated and tested with the documentation build system
- [x] The feature has been verified to work with the conventional build system
- [ ] The feature has been verified to work with the CMake based build system
- [ ] Suitable tests have been added to the unittest tree.
- [ ] One or more example input decks are included

**Further Information, Files, and Links**