@akohlmey 2022-07-08
FYI, there are three segmentation faults for unit tests. This often happens when there is some uninitialized data.