@stanmoore1 2022-07-11
This is working in all my tests. One question, should `neigh/trim yes` be the default since it is normally faster, or should we keep the old behavior? @sjplimp @akohlmey