@akohlmey 2022-07-11
> This is working in all my tests. One question, should `neigh/trim yes` be the default since it is normally faster, or should we keep the old behavior? @sjplimp @akohlmey 

There still seem to be test failures with the embedded unit tests. I will need some time to reconfigure my phone to find a provider that supports roaming and the wifi hotspot of my cell phone before I can check it out in detail.

If we find and resolve the causes of the failures, I would suggest to make the new option the default. We are at the beginning of a development cycle and now is the best time for such changes and this way we increase the chances of getting bug reports.
If it turns out to be too risky, we can always roll it back and put a warning into the manual (faster but may not work for all cases).