@akohlmey 2022-07-12
@stanmoore1 I fixed two bugs, one due to an uninitialized variable and one in pair style hybrid/scaled due to an unallocated array (that style overrides the settings() functions from pair style hybrid).