@akohlmey 2022-07-12
There are still issues running with OpenMP, e.g. the command `~/compile/lammps/build-test/lmp -in in.snap.InP.JCPA2020 -sf omp` in the examples/snap folder yields:
```
Hybrid pair style last /omp style zbl
Last active /omp style is pair_style zbl
ERROR: Requested neighbor pair method does not exist (src/neighbor.cpp:928)
Last command: run             ${nsteps}
```
Please see: https://ci.lammps.org/job/dev/job/pull_requests/job/ubuntu/job/run_tests/job/cmake_testing_openmp/563/
There may be similar issues when using the INTEL package as well.