@stanmoore1 2022-07-12
@akohlmey thanks for working on this--I will fix the OMP/INTEL issue