@akohlmey 2022-07-12
@stanmoore1 wouldn't it be cleaner to just define "forwarders" to the non-accelerated version like for the "skip" neighbor list classes in `src/OPENMP/npair_skip_omp.h`?