@akohlmey 2022-07-12
> I don't think that will work for INTEL though, since the structure of the neigh list is different?

I don't know. Haven't looked at those in a long time. That would be a question for @wmbrownIntel