@stanmoore1 2022-07-12
The Intel trim is broken in my test, debugging...