@stanmoore1 2022-07-12
@akohlmey this is now passing my tests. Besides the issue of `pair_hybrid` not handling a pair_style that has multiple neigh list requests, this is ready to merge from my POV.