@akohlmey 2022-07-12
> One other issue I noticed is that this code in `pair_hybrid` doesn't work if a pair_style has multiple neigh list requests (e.g. MEAM): https://github.com/lammps/lammps/blob/develop/src/pair_hybrid.cpp#L622, since it just finds the first request and then breaks out.
> 

This sounds like a job for @sjplimp to look into, and it may be a more general problem that could be reported as a separate issue and then addressed in a separate pull request.